requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
google-generativeai
google-cloud-documentai
google-cloud-speech
google-api-core
python-jose